MOD/#5: 스크린별 Destination 객체 생성 위치 변경
