+++ app/src/main/java/com/sopt/now/compose/ui/SoptApp.kt
@@ -1,32 +1,24 @@
 package com.sopt.now.compose.ui
 
-import android.content.Context
-import android.widget.Toast
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.Icon
 import androidx.compose.material3.NavigationBar
 import androidx.compose.material3.NavigationBarItem
-import androidx.compose.material3.NavigationBarItemDefaults
-import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Surface
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.unit.dp
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.currentBackStackEntryAsState
 import androidx.navigation.compose.rememberNavController
-import com.sopt.now.compose.ui.screens.home.HomeDestination
 import com.sopt.now.compose.ui.navigation.ApplicationNavHost
+import com.sopt.now.compose.ui.navigation.HomeDestination
 import com.sopt.now.compose.ui.navigation.NavigationDestination
-import com.sopt.now.compose.ui.screens.profile.ProfileDestination
-import com.sopt.now.compose.ui.screens.search.SearchDestination
+import com.sopt.now.compose.ui.navigation.ProfileDestination
+import com.sopt.now.compose.ui.navigation.SearchDestination
 
 @Composable
 fun SoptApp() {
